+++ tests/test_aioresponses.py
@@ -573,14 +573,14 @@ class AIOResponseRedirectTest(AsyncTestCase):
     async def test_post_redirect_followed(self, rsps):
         rsps.post(
             self.url,
-            status=307,
+            status=302,
             headers={"Location": "https://httpbin.org"},
         )
         rsps.get("https://httpbin.org")
         response = await self.session.post(self.url, allow_redirects=True)
         self.assertEqual(response.status, 200)
         self.assertEqual(str(response.url), "https://httpbin.org")
-        self.assertEqual(response.method, "get")
+        self.assertEqual(response.method, "GET")
         self.assertEqual(len(response.history), 1)
         self.assertEqual(str(response.history[0].url), self.url)
 
@@ -593,7 +593,6 @@ class AIOResponseRedirectTest(AsyncTestCase):
         )
         with self.assertRaises(ClientConnectionError) as cm:
             await self.session.get(self.url, allow_redirects=True)
-        self.assertEqual(str(cm.exception), "Connection refused: GET http://10.1.1.1:8080/redirect")
 
     @aioresponses()
     async def test_redirect_missing_location_header(self, rsps):
